fix(server): no project commands would fail if using secrets
