+++ core/src/template-string/template-string.ts
@@ -203,9 +203,7 @@ export function resolveTemplateString(string: string, context: ConfigContext, op
     const prefix = `Invalid template string (${truncate(string, 35).replace(/\n/g, "\\n")}): `
     const message = err.message.startsWith(prefix) ? err.message : prefix + err.message
 
-    throw new TemplateStringError(message, {
-      err,
-    })
+    throw new TemplateStringError(message, {})
   }
 }
