+++ core/test/unit/src/commands/create/create-module.ts
@@ -9,8 +9,7 @@
 import { expect } from "chai"
 import { withDefaultGlobalOpts, TempDirectory, makeTempDir, expectError } from "../../../../helpers"
 import { CreateModuleCommand, getModuleTypeSuggestions } from "../../../../../src/commands/create/create-module"
-import { makeDummyGarden } from "../../../../../src/cli/cli"
-import { Garden } from "../../../../../src/garden"
+import { Garden, makeDummyGarden } from "../../../../../src/garden"
 import { join } from "path"
 import { pathExists, readFile, writeFile, mkdirp } from "fs-extra"
 import { safeLoadAll } from "js-yaml"
